+++ src/components/common/ProgressBar/ProgressBar.tsx
@@ -1,4 +1,4 @@
-import { useColorContext } from '#components/pages/Providers/ColorProvider'
+import { useColorContext } from '#components/templates/Providers/ColorProvider'
 import React, { useEffect, useState } from 'react'
 
 export default function ProgressBar() {
