+++ src/components/templates/Providers/ColorProvider/index.tsx
